+++ gesture_detection/gesture_recognition.py
@@ -5,6 +5,7 @@ from datetime import datetime, date
 import matplotlib.pyplot as plt
 import json
 import sys
+import os
 
 
 def get_model():
@@ -98,21 +99,22 @@ model.load_weights("Train_weights_1.h5")
 # c7- talking to passenger
 classes = ["safe_driving", "texting", "talking_on_phone", "operating_center_console", "drinking", "reaching_behind",
            "hair_makeup", "talking_to_passenger"]
+session_date = date.today().strftime("%d/%m/%Y")
+stop_at = 600
 data_dict = {
-    "session_date": date.today().strftime("%d/%m/%Y"),
-    "session_time": datetime.now().strftime("%H:%M:%S"),
-    "data": {
-        "timeline": []
-    },
-    "development": {
-        "safe_driving": [],
-        "texting": [],
-        "talking_on_phone": [],
-        "operating_center_console": [],
-        "drinking": [],
-        "reaching_behind": [],
-        "hair_makeup": [],
-        "talking_to_passenger": []
+    "session_date": session_date,
+    "gesture_data": {
+        "timeline": [],
+        "development": {
+            "safe_driving": [],
+            "texting": [],
+            "talking_on_phone": [],
+            "operating_center_console": [],
+            "drinking": [],
+            "reaching_behind": [],
+            "hair_makeup": [],
+            "talking_to_passenger": []
+        }
     }
 }
 
@@ -177,14 +179,15 @@ while True:
     # Add the detected gesture into the timeline.
     for gesture_type in classes:
         if gesture_type == label:
-            data_dict["development"][f"{label}"].append(1)
+            data_dict["gesture_data"]["development"][f"{label}"].append(1)
         else:
-            data_dict["development"][f"{gesture_type}"].append(0)
+            data_dict["gesture_data"]["development"][f"{gesture_type}"].append(0)
 
-    resized_img = cv2.resize(test_img, (1000, 700))
-    cv2.imshow('Gesture Analysis', resized_img)
+    cv2.imshow('Gesture Analysis', test_img)
     key = cv2.waitKey(1) & 0xFF
     if key == ord('q'):
+        break
+    if current_frame == stop_at:
         break
 cv2.destroyAllWindows()
 cap.release()
@@ -193,11 +196,11 @@ intervals_total = (current_frame / (frame_rate*capture_interval))
 
 print(f"total captured frames: {current_frame}")
 print(f"total captured intervals: {intervals_total}")
-for i, item in enumerate(data_dict["development"]):
-    print(f"{item} total data points: {len(data_dict['development'][item])}")
+for i, item in enumerate(data_dict["gesture_data"]["development"]):
+    print(f"{item} total data points: {len(data_dict['gesture_data']['development'][item])}")
 
 for gesture_type in classes:
-    print(f"""{gesture_type} detections: {data_dict['development'][f'{gesture_type}'].count(1)}""")
+    print(f"""{gesture_type} detections: {data_dict['gesture_data']['development'][f'{gesture_type}'].count(1)}""")
 
 
 # Record in JSON file the data captured in intervals so there aren't too many data points per minute.
@@ -207,22 +210,30 @@ for i in range(current_frame):
     if i % (frame_rate*capture_interval) == 0:
         added_data = False
         for gesture_type in classes:
-            if data_dict['development'][f'{gesture_type}'][i] == 1:
-                data_dict['data']['timeline'].append(classes.index(gesture_type))
+            if data_dict['gesture_data']['development'][f'{gesture_type}'][i] == 1:
+                data_dict['gesture_data']['timeline'].append(classes.index(gesture_type))
                 added_data = True
         if not added_data:
-            data_dict['data']['timeline'].append(-1)
+            data_dict['gesture_data']['timeline'].append(-1)
 # Example output:
 # timeline: [-1, -1, -1, 0, 0, 1, 1, 5, 5, -1, -1, 0, ...]
 
-with open('gesture_result.json', 'w') as json_file:
-    json.dump(data_dict, json_file)
 
+if not os.path.exists("gesture_result.json"):
+    with open('gesture_result.json', 'w') as json_file:
+        json.dump({"sessions": [data_dict]}, json_file)
+else:
+    with open('gesture_result.json', 'w+') as json_file:
+        json_data = json.load(json_file)
+        for i, item in enumerate(json_data["sessions"]):
+            if json_data["sessions"][i]["session_date"] == session_date:
+                json_data["sessions"][i] = data_dict
+        json.dump(json_data, json_file)
 
 class_counts = [i for i, _ in enumerate(classes)]
 gesture_counts = []
 for gesture_type in classes:
-    gesture_counts.append(data_dict['development'][f'{gesture_type}'].count(1))
+    gesture_counts.append(data_dict['gesture_data']['development'][f'{gesture_type}'].count(1))
 
 plt.style.use('ggplot')
 plt.bar(class_counts, gesture_counts, color='blue')
